+++ pkg/reconciler/actions.go
@@ -1,4 +1,4 @@
-package reconcileaction
+package reconciler
 
 import (
 	"fmt"
