+++ pkg/reconciler/procedures_test.go
@@ -1,4 +1,4 @@
-package reconcileaction
+package reconciler
 
 import (
 	"testing"
